fix: remove profile_pic_url field on image deletion

+++ controllers/image/deleteImg.js
@@ -1,11 +1,18 @@
+const User = require('../../models/users');
 const deleteImgFromCloud = require('../../utils/image/delImg');
-const { NotFoundErr, ForbiddenErr } = require('../../utils/errors/errors');
+const { ForbiddenErr } = require('../../utils/errors/errors');
 
 const controller = async (req, res, next) => {
   let imageAuthor = req.query.name.slice(0, req.query.name.indexOf('uid'));
 
   if (imageAuthor !== req.user._id.toString()) {
     return next(new ForbiddenErr());
+  }
+
+  if (req.query.name === req.user.profile_pic_url) {
+    let user = await User.findById(req.user._id).select('profile_pic_url').catch(next);
+    user.profile_pic_url = null;
+    user.save();
   }
 
   await deleteImgFromCloud(req.query.name).catch(next);
